Bugfix: Don't delete parent Group in H5Group class
The H5Group delete() method would delete the "self" group after the
deletion of its last child.  The purpose of this was to remove empty
container groups such as "data_arrays" and "multi_tags".  The bug was
that this would also delete any H5Group that has no children once its
last child was removed.  For example, a Block would be removed if its
metadata was removed and it had no other subgroups (no sub-objects had
been created so far).